# seen tracker: use array for storage